the same but using a double: behavior is different

+++ src/ch6/arithmetic/operators.cpp
@@ -3,7 +3,7 @@
 int main() {
   constexpr int apples {12};
   std::cout << "You have " << apples << " apples. Enter how many people to divide them between: ";
-  int people{};
+  double people{};
   std::cin >> people;
   std::cout << "Each person gets " << apples / people << " apples.\n";
   return 0;
